booking_service/requirements.txt
fastapi
uvicorn[standard]
kafka-python==2.0.2
pymongo
requests
networkx
pytest